% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{agreement}
\alias{agreement}
\title{Calculate agreement between two clustering results}
\usage{
agreement(clst1, clst2)
}
\arguments{
\item{clst1}{First clustering result}

\item{clst2}{Second clustering result}
}
\value{
Vector of agreement values
}
\description{
Calculate agreement between two clustering results
}
