% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{agreement_nk}
\alias{agreement_nk}
\title{Calculate agreement between two clustering results with known number of clusters}
\usage{
agreement_nk(clst1, clst2, nk)
}
\arguments{
\item{clst1}{First clustering result}

\item{clst2}{Second clustering result}

\item{nk}{Number of clusters}
}
\value{
Vector of agreement values
}
\description{
Calculate agreement between two clustering results with known number of clusters
}
