% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cluster_utils.R
\name{calculate_comparison_stats}
\alias{calculate_comparison_stats}
\title{Calculate Comparison Statistics}
\usage{
calculate_comparison_stats(results, method_names)
}
\arguments{
\item{results}{List of results for each method}

\item{method_names}{Names of combination methods}
}
\value{
List of comparison statistics
}
\description{
Calculate Comparison Statistics
}
\keyword{internal}
