% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cluster_utils.R
\name{calculate_stability_measures}
\alias{calculate_stability_measures}
\title{Calculate Stability Measures for a Clustering Method}
\usage{
calculate_stability_measures(
  x,
  k,
  scheme,
  B = 100,
  n_ref = 3,
  hc.method = "ward.D",
  dist_method = "euclidean"
)
}
\arguments{
\item{x}{Input data matrix}

\item{k}{Number of clusters}

\item{scheme}{Clustering scheme ("kmeans", "hc", or "spectral")}

\item{B}{Number of bootstrap samples}

\item{n_ref}{Number of reference distributions}

\item{hc.method}{Hierarchical clustering method}

\item{dist_method}{Distance method}
}
\value{
List containing stability measures and clustering results
}
\description{
Calculate Stability Measures for a Clustering Method
}
\keyword{internal}
