% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_analysis.R
\name{compare_moc_results}
\alias{compare_moc_results}
\title{Compare MOC Results}
\usage{
compare_moc_results(results, metric = "ari", plot = TRUE)
}
\arguments{
\item{results}{Results from analyze_moc_datasets}

\item{metric}{Metric to compare ("ari", "runtime", or "modularity")}

\item{plot}{Whether to generate comparison plot (default: TRUE)}
}
\description{
Compare MOC Results
}
