% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cluster_utils.R
\name{create_graph_and_communities}
\alias{create_graph_and_communities}
\title{Create Graph and Find Communities Using Different Methods}
\usage{
create_graph_and_communities(inc, method = "fastgreedy")
}
\arguments{
\item{inc}{Incidence matrix}

\item{method}{Community detection method: "fastgreedy", "metis", "hmetis", or "all" (default: "fastgreedy")}
}
\value{
List containing graph and community detection results from specified method(s)
}
\description{
Create Graph and Find Communities Using Different Methods
}
\keyword{internal}
