% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cluster_utils.R
\name{create_incidence_matrix}
\alias{create_incidence_matrix}
\title{Create Incidence Matrix for Graph Construction}
\usage{
create_incidence_matrix(
  results_km,
  results_hc,
  results_sc,
  k_km,
  k_hc,
  k_sc,
  Smin_delta_km,
  Smin_delta_hc,
  Smin_delta_sc,
  combine_fn,
  n_samples
)
}
\arguments{
\item{results_km}{K-means results}

\item{results_hc}{Hierarchical clustering results}

\item{results_sc}{Spectral clustering results}

\item{k_km}{Number of k-means clusters}

\item{k_hc}{Number of hierarchical clusters}

\item{k_sc}{Number of spectral clusters}

\item{Smin_delta_km}{K-means stability delta}

\item{Smin_delta_hc}{Hierarchical clustering stability delta}

\item{Smin_delta_sc}{Spectral clustering stability delta}

\item{combine_fn}{Function to combine stability measures}

\item{n_samples}{Number of samples}
}
\value{
Incidence matrix for graph construction
}
\description{
Create Incidence Matrix for Graph Construction
}
\keyword{internal}
