% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cluster_utils.R
\name{define_combination_methods}
\alias{define_combination_methods}
\title{Define Stability Combination Methods}
\usage{
define_combination_methods(alpha = 0.5)
}
\arguments{
\item{alpha}{Weight for weighted combination}
}
\value{
List of combination functions
}
\description{
Define Stability Combination Methods
}
\keyword{internal}
