% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{min_agreement}
\alias{min_agreement}
\title{Calculate minimum agreement across clusters}
\usage{
min_agreement(clst, agrmt)
}
\arguments{
\item{clst}{clustering result vector}

\item{agrmt}{agreement values vector}
}
\value{
minimum average agreement value across clusters
}
\description{
Calculates the minimum average agreement value across all clusters
}
