% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_analysis.R
\name{plot_moc_results}
\alias{plot_moc_results}
\title{Plot MOC Results}
\usage{
plot_moc_results(
  results,
  dataset_names,
  methods = c("kmeans", "hierarchical", "spectral", "fastgreedy", "metis", "hmetis"),
  plot_file = NULL,
  max_plots_per_page = 12,
  mar = c(2, 2, 2, 1)
)
}
\arguments{
\item{results}{Results from analyze_moc_datasets}

\item{dataset_names}{Name or vector of names of datasets to plot}

\item{methods}{Methods to plot (default: c("fastgreedy", "metis", "hmetis"))}

\item{plot_file}{Output file for plots (default: NULL)}

\item{max_plots_per_page}{Maximum number of plots per page (default: 12)}

\item{mar}{Margins for plots (default: c(2, 2, 2, 1))}
}
\description{
Plot MOC Results
}
