% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_dist.R
\name{ref_dist}
\alias{ref_dist}
\title{Generate reference distribution for stability assessment}
\usage{
ref_dist(df)
}
\arguments{
\item{df}{data.frame or matrix of the original dataset}
}
\value{
A scaled matrix containing the reference distribution
}
\description{
Generates a reference distribution by sampling from uniform distributions
with ranges determined by the original data.
}
\details{
Generate Reference Distribution
}
\examples{
\donttest{
data(iris)
df <- iris[,1:4]
ref <- ref_dist(df)
}

}
