% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_dist.R
\name{ref_dist_bin}
\alias{ref_dist_bin}
\title{Generate reference distribution for binary data}
\usage{
ref_dist_bin(df)
}
\arguments{
\item{df}{data.frame or matrix of the original binary dataset}
}
\value{
A matrix containing the permuted binary reference distribution
}
\description{
Generates a reference distribution by randomly permuting each column
of the original binary dataset.
}
\details{
Generate Binary Reference Distribution
}
\examples{
\donttest{
binary_data <- matrix(sample(0:1, 100, replace=TRUE), ncol=5)
ref <- ref_dist_bin(binary_data)
}

}
