% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_dist.R
\name{ref_dist_pca}
\alias{ref_dist_pca}
\title{Generate PCA-based reference distribution}
\usage{
ref_dist_pca(df)
}
\arguments{
\item{df}{data.frame or matrix of the original dataset}
}
\value{
A scaled matrix containing the reference distribution in PCA space
}
\description{
Generates a reference distribution in PCA space by sampling from uniform
distributions with ranges determined by the PCA-transformed data.
}
\details{
Generate Reference Distribution using PCA
}
\examples{
\donttest{
data(iris)
df <- iris[,1:4]
ref <- ref_dist_pca(df)
}

}
