% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_sub.R
\name{scheme_sub_hc}
\alias{scheme_sub_hc}
\title{Subsampling-based Hierarchical Clustering}
\usage{
scheme_sub_hc(
  df,
  nk,
  B = 10,
  cut_ratio = 0.5,
  hc.method = "ward.D",
  dist_method = "euclidean"
)
}
\arguments{
\item{df}{data.frame or matrix of the dataset}

\item{nk}{number of clusters}

\item{B}{number of bootstrap samples}

\item{cut_ratio}{ratio for subsampling (default: 0.5)}

\item{hc.method}{hierarchical clustering method (default: "ward.D")}

\item{dist_method}{distance method (default: "euclidean")}
}
\value{
List containing clustering results and stability measures
}
\description{
Subsampling-based Hierarchical Clustering
}
\keyword{internal}
