% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_sub.R
\name{scheme_sub_km}
\alias{scheme_sub_km}
\title{Subsampling-based K-means Clustering}
\usage{
scheme_sub_km(df, nk, B = 10, cut_ratio = 0.5, r = 5)
}
\arguments{
\item{df}{data.frame or matrix of the dataset}

\item{nk}{number of clusters}

\item{B}{number of bootstrap samples}

\item{cut_ratio}{ratio for subsampling (default: 0.5)}

\item{r}{number of k-means runs (default: 5)}
}
\value{
List containing clustering results and stability measures
}
\description{
Subsampling-based K-means Clustering
}
\keyword{internal}
