% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Data Set}
\format{
The data set \code{wine} contains a \code{data.frame} of 14 variables. The first variable is the types
of wines. The other 13 variables are quantities of the constituents.
}
\usage{
data(wine)
}
\description{
These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from
three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three
types of wines.
}
\references{
https://archive.ics.uci.edu/ml/datasets/wine
}
